+++ x/genaccounts/internal/types/expected_keepers.go
@@ -1,4 +1,4 @@
-package genaccounts
+package types
 
 import (
 	sdk "github.com/cosmos/cosmos-sdk/types"
