+++ x/genaccounts/internal/types/genesis_account_test.go
@@ -1,4 +1,4 @@
-package genaccounts
+package types
 
 import (
 	"errors"
