+++ x/genaccounts/internal/types/genesis_state.go
@@ -1,4 +1,4 @@
-package genaccounts
+package types
 
 import (
 	"encoding/json"
