+++ x/genaccounts/internal/types/genesis_state_test.go
@@ -1,4 +1,4 @@
-package genaccounts
+package types
 
 import (
 	"testing"
